fix memory corruption
In order to increase C2P speed on 040/060 processors,
DoomAttack 'massages' the  cache settings for the chipram that
the screen's bitplanes reside in.
To do so, the bitplanes need to be page aligned.
This means, when allocating the bitplanes, we need to be able
to push them further out onto the next 4kb boundary.

The bitplanes for native modes get allocated as just one raster
with 8x the height; DoomAttack places the indiviual bitplanes
in serial into that raster. To get the page-alignment working,
up to 4095 bytes more need to be allocated. We achieve that by
adding 'spare height' to the raster.

The original formula did not properly calculate the amount of
necessary spare height, thus the 4kb alignment would push
the bitplanes too far out, causing memory corruption.