Move plugin startup code into .text segment
This will ensure, the startup code will end up first in the
object file and thus DoomAttack will identify the files as
plugins.

Currently, only the C2P plugin would actually walk the
seqment list, if the startup code was not found
int the first segment. And even that code was broken.