Create runtime environment for plugins
Basically, create a form of startup code for runtime
loaded plugins, so they can use code like printf, scanf etc.

This may bloat the plugins a little. An alternative would be
to provide pointers to std functions to the plugins.

On unresolved problem: why is __init_libs not automatically
included in the __INIT_LIST__?