Work around linker issue with __INIT_LIST__ and __EXIT_LIST__
Some plugins may want to use libnix functionality like printf.
The plugins are loaded via LoadSeg and use a custom startup code linked with -nostartfiles.
This presents a problem as we need to emulate the way libnix is initialized.

We do this by manually traversing the __INIT_LIST__ and __EXIT_LIST__.

A recent toolchain change requires a change to the way these symbols get introduced
to the linker.