Fix Midiplugin
The assembly code is using DOSBase to open the music instrument file.
The C compiler/linker didn't notice that DOSBase was used and thus
didn't generate a stub to auto-initialize it.

But, the C_DAM_Init function gets supplied with DOSBase anyway,
so take it from there.

I think this is intended/best way anyways.